[Web] Refactor: :rainbow: separate NicknameInput into a molecule and an organism components #364
NicknameInput을 molecule과 organism으로 분리
